parallel stddev
