+++ core/src/test/resources/sqllogictest/test/sql/test_stddev.test
@@ -1,0 +1,134 @@
+statement ok
+create table stddev_test(val int, grp int)
+
+statement ok
+insert into stddev_test values (42, 1), (43, 1), (42, 2), (1000, 2), (NULL, 1), (NULL, 3)
+
+query I
+SELECT stddev_samp(1)
+----
+NULL
+
+query I
+SELECT var_samp(1)
+----
+NULL
+
+# stddev_samp
+query R
+select round(stddev_samp(val), 1) from stddev_test
+----
+478.800000
+
+query R
+select round(stddev_samp(val), 1) from stddev_test  where val is not null
+----
+478.800000
+
+query IRRI
+select grp, sum(val), round(stddev_samp(val), 1), min(val) from stddev_test group by grp order by grp
+----
+1 85 0.700000 42
+2 1042 677.400000 42
+3 NULL NULL NULL
+
+query IRRI
+select grp, sum(val), round(stddev_samp(val), 1), min(val) from stddev_test where val is not null group by grp order by grp
+----
+1 85 0.700000 42
+2 1042 677.400000 42
+
+# stddev_pop
+query R
+select round(stddev_pop(val), 1) from stddev_test
+----
+414.700000
+
+query R
+select round(stddev_pop(val), 1) from stddev_test  where val is not null
+----
+414.700000
+
+query IRRI
+select grp, sum(val), round(stddev_pop(val), 1), min(val) from stddev_test group by grp order by grp
+----
+1 85 0.500000 42
+2 1042 479.000000 42
+3 NULL NULL NULL
+
+query IRRI
+select grp, sum(val), round(stddev_pop(val), 1), min(val) from stddev_test where val is not null group by grp order by grp
+----
+1 85 0.500000 42
+2 1042 479.000000 42
+
+# var_samp
+query R
+select round(var_samp(val), 1) from stddev_test
+----
+2.292816e+5
+
+query R
+select round(variance(val), 1) from stddev_test
+----
+2.292816e+5
+
+query R
+select round(var_samp(val), 1) from stddev_test where val is not null
+----
+2.292816e+5
+
+query IRRI
+select grp, sum(val), round(var_samp(val), 1), min(val) from stddev_test group by grp order by grp
+----
+1 85 0.500000 42
+2 1042 4.58882e+5 42
+3 NULL NULL	NULL
+
+query IRRI
+select grp, sum(val), round(var_samp(val), 1), min(val) from stddev_test where val is not null group by grp order by grp
+----
+1 85 0.500000 42
+2 1042 4.58882e+5 42
+
+# var_pop
+query R
+select round(var_pop(val), 1) from stddev_test
+----
+1.719612e+5
+
+query R
+select round(var_pop(val), 1) from stddev_test  where val is not null
+----
+1.719612e+5
+
+query IRRI
+select grp, sum(val), round(var_pop(val), 2), min(val) from stddev_test group by grp order by grp
+----
+1 85 0.250000 42
+2 1042 2.29441e+5 42
+3 NULL NULL NULL
+
+query IRRI
+select grp, sum(val), round(var_pop(val), 2), min(val) from stddev_test where val is not null group by grp order by grp
+----
+1 85 0.250000 42
+2 1042 2.29441e+5 42
+
+statement ok
+create table stddev_test_alias(val int, grp int)
+
+statement ok
+insert into stddev_test_alias values (42, 1), (43, 1), (42, 2), (1000, 2), (NULL, 1), (NULL, 3)
+
+# stddev_samp
+query R
+select round(stddev(val), 1) from stddev_test_alias
+----
+478.800000
+
+query I
+select stddev(0) from long_sequence(10)
+----
+0.000000
+
